scripts/requirements.txt
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
supabase>=2.0.0
